requirements.txt
flask
requests
JPype1 >= 1.4.1